chore:delete route general-details and add page principal
