+++ src/app/dashboard/settings/integrations/page.tsx
@@ -1,3 +1,8 @@
+export const metadata = {
+    title: "Integraciones",
+};
+
+
 export default function Page() {
-  return <>Integracion</>;
+    return <>Integracion</>;
 }
