+++ src/app/dashboard/settings/notifications/page.tsx
@@ -1,3 +1,9 @@
+export const metadata = {
+    title: "Notificaciones",
+};
+
+
+
 export default function Page() {
-  return <>Notificacion</>;
+    return <>Notificacion</>;
 }
