+++ src/app/dashboard/settings/security/page.tsx
@@ -1,3 +1,7 @@
+export const metadata = {
+    title: "Seguridad",
+};
+
 export default function Page() {
-  return <>Seguridad</>;
+    return <>Seguridad</>;
 }
